requirements.txt
python-dotenv
requests
groq
pyfiglet
termcolor
beautifulsoup4

